requirements.txt
asgiref==3.5.2
Django==4.1.3
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.0
phonenumberslite==8.13.0
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3